🐛 FIX: Payment 모델의 amount 필드 타입 변경

+++ payments/models.py
@@ -65,7 +65,7 @@ class Payment(models.Model):
     course = models.ForeignKey(
         Course, on_delete=models.CASCADE, related_name="payments"
     )
-    amount = models.DecimalField(max_digits=10, decimal_places=0, help_text="결제 금액")
+    amount = models.PositiveIntegerField(help_text="결제 금액")
     payment_method = models.CharField(
         max_length=20, choices=PAYMENT_METHOD_CHOICES, null=True, blank=True
     )
